bug: do not parse issuer as url because issuer must be exact
some providers have a "/" at the end of the issuer in the openid config and some do not. we previously stripped any "/" characters away, but Auth0 is now signing tokens with a slash. in that case, the issuer was parsed with a slash that we then stirpped away resulting in a issuer mismatch.

Signed-off-by: Anton Engelhardt <antoncengelhardt@icloud.com>